fix. External Api GET, on AccountService, StockService, Json response

+++ src/main/java/com/investing/api/entity/dto/StockAccountResponseDto.java
@@ -9,6 +9,6 @@ public record StockAccountResponseDto(UUID account_id,
                                       String currency,
                                       String shortName,
                                       Long quantity,
-                                      Double actualRegularMarketPrice,
-                                      Double totalValue) {
+                                      BigDecimal actualRegularMarketPrice,
+                                      BigDecimal totalValue) {
 }
